[doc] add internals.md
This document is an attempt to describe the internals of LK for ARM
architecture. It is aimed at people who are new to LK source. In some parts,
detailed explanation is given regarding to the architecture and instructions so
that people who are new to ARM assembly are not get lost.