[make] Cleanup overlay processing order
Add LKROOT to LKINC only if it is not already there.
This would allow to specify particular overlay processing
order if required.

Remove adding separate LKROOT/include to GLOBAL_DEFINES as
It is already a part of LKINC list.

Add all directories specified by LKINC to GLOBAL_LDFLAGS instead
of just LKROOT

Change-Id: Ia61623f88f2c5978182c40b09d98b6e5b84e3955