[kernel] thread_join from multiple threads
This patch enables multiple concurrent threads to call thread_join on
a particular thread

In thread_join(), we simply make checks regarding whether the thread
structure has been freed or not.

This could just be a quick fix. A better solution would be to
implement reference counting on thread objects, and free the
resources when the count goes to zero.