[app][mdebug] avoid getting stuck if IN ACK is lost
If the host receives the IN response packet but the ACK from the host
is lost, we'd be stuck until the next IN, which wouldn't come because
the host does a simple send-request / read-response thing.  Instead,
don't wait for IN txns to complete unless there's already one outstanding.

This allows us to read the next request, at which point the host will
start an IN txn for the response, which will cause the previous IN txn
to complete.