[external] update raspberry pi pico SDK to 2.1.0
Mostly a verbatim copy of what is inside the src dir of the pico-sdk at
https://github.com/raspberrypi/pico-sdk.

Tweaked
rp2_common/pico_platform_compiler/include/pico/platform/compiler.h
to always insert an inline inside __always_inline which seems to be
required for GCC 14.

Updated platform/rp20xx/rules.mk to include new pieces of the SDK that
are apparently required for booting.