[includes] remove lk/utils.h which seem to be unused on anything in mainline
Was a bit of a hodgepodge of defines. If any of these are needed we can
add them back somewhere a bit more appropriate.

ARRAY_SIZE can use countof() defined in compiler.h

KB/MB/GB are generally duplicately defined in architecture specific
files, but not generically. I couldn't think off the top of my head of a
better place to generically put these, especially since they're used in
asm files, so just deleted it for now.

The other stuff seemed to be just unused anywhere.