[rust] Add initial Rust support
This brings in basic support for building Rust code as part of LK. The
build system compiles a single Rust crate (using Cargo) into a static
library.

The crate is a template, to deal with the assumption Cargo makes about
the top-level crate directing the configuration of the entire build.  As
part of the build, these files are copied into the build directory, as
well as a `.cargo/config.toml` file that sets up the compilation for
this build with the rust compiler.

By doing it this way, it is simply a matter of pointing a tool, such as
vscode, to this directory, and the rust-analyzer will work as expected.

Signed-off-by: David Brown <david.brown@linaro.org>